+++ js/symbol/glyph_source.js
@@ -28,7 +28,7 @@ GlyphSource.prototype.getSimpleGlyphs = function(fontstack, glyphIDs, uid, callb
         this.stacks[fontstack] = {};
     }
     if (this.atlases[fontstack] === undefined) {
-        this.atlases[fontstack] = new GlyphAtlas(128, 128);
+        this.atlases[fontstack] = new GlyphAtlas();
     }
 
     var glyphs = {};
@@ -106,7 +106,7 @@ GlyphSource.prototype.loadRange = function(fontstack, range, callback) {
         var url = glyphUrl(fontstack, rangeName, this.url);
 
         getArrayBuffer(url, function(err, data) {
-            var glyphs = !err && new Glyphs(new Protobuf(new Uint8Array(data)));
+            var glyphs = !err && new Glyphs(new Protobuf(data));
             for (var i = 0; i < loading[range].length; i++) {
                 loading[range][i](err, range, glyphs);
             }